Made ProjectManager abstrsact.  Moved SignalR integration into DashboardProjectManager.
